+++ web/documentserver-example/go/server/api/default/config.go
@@ -22,10 +22,12 @@ import (
 	"fmt"
 	"net/http"
 	"strings"
+	"time"
 
 	"github.com/ONLYOFFICE/document-server-integration/server/managers"
 	"github.com/ONLYOFFICE/document-server-integration/server/models"
 	"github.com/ONLYOFFICE/document-server-integration/server/shared"
+	"github.com/golang-jwt/jwt"
 )
 
 func (srv *DefaultServerEndpointsHandler) Config(w http.ResponseWriter, r *http.Request) {
@@ -70,6 +72,10 @@ func (srv *DefaultServerEndpointsHandler) Config(w http.ResponseWriter, r *http.
 				remoteAddr,
 			),
 			Mode: "edit",
+		},
+		StandardClaims: jwt.StandardClaims{
+			ExpiresAt: time.Now().Add(time.Minute * srv.config.JwtExpiresIn).Unix(),
+			IssuedAt:  time.Now().Unix(),
 		},
 	}